+++ src/app/core/services/index.ts
@@ -4,3 +4,5 @@ export * from './local-storage.service';
 export * from './generator.service';
 export * from './user-profile.service';
 export * from './dialog.service';
+export * from './app-settings.service';
+export * from './api-config.service';